+++ inc/lilyNotation/index.ts
@@ -3,6 +3,7 @@ import {Note, Notation, assignNotationEventsIds} from "./notation";
 import {fuzzyMatchNotations} from "./fuzzyMatch";
 import {matchWithMIDI, matchWithExactMIDI} from "./matcher";
 import Scheduler from "./scheduler";
+import ImplicitType from "./implicitType";
 import * as measureLayout from "../measureLayout";
 import {LayoutType} from "../measureLayout";
 
@@ -20,6 +21,7 @@ export {
 	matchWithMIDI,
 	matchWithExactMIDI,
 	Scheduler,
+	ImplicitType,
 	LayoutType,
 	MLayoutClasses,
 };